Create BadiRate setup table and corresponding scripts for PBS job arrays

Create BadiRate setup table and corresponding scripts for PBS job arrays

prepare_badirate(og_path, tree, branch_models = c(gr = "GR", fr = "FR",
  sp = "15->16"), rate_model = "GD", estimation = "ML",
  out_dir = "./raw_outputs", script_dir = "./scripts",
  replicates = 2, ancestral = TRUE, outlier = TRUE,
  seed = 20180808, start_value = 1, pbs_q = "smps",
  badirate_path = "", create_scripts = "pbs")

Arguments

og_path
Path to single-file orthogroup folder, created with `split_orthogroups()`

tree
Path to ultrametric species tree file. Leaf names have to match to orthogroup header names.

branch_models
Vector of named branch models. Name should be a 2-letter abbreviation of choice (except gr and fr, which are specific). Value should be either "GR", "FR" or a specific branch model in the syntax of badirate (e.g. "13->14:13->15_18->21"). *CAUTION: More complex models can increase running time drastically.*

rate_model
Rate model. Deafult is GD.

estimation
Estimation method used for BadiRate. Default is ML.

out_dir
Output directory for the future runs.

script_dir
Directory which the script files should be saved.

replicates
How many replications per model. Default is 2. *CAUTION: Big number of replicates can drastically extend runtime*

ancestral
Logical. See BadiRate -anc option.

outlier
Logical. See BadiRate -outlier option.

seed
Initial seed value to set.

start_value
0 or 1. 0 will start parameter search from Maximum Parsimony, 1 will choose a random number.

pbs_q
Queue of PBS. Default is smps.

badirate_path
Path to BadiRate.pl executabe if not in $PATH.

create_scripts
Logical. Create pbs job scripts if TRUE, which is the default.

Value

Set up table of experimetns. Also writes out script files.